backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9
django-cors-headers>=3.13
python-dotenv>=1.0
gunicorn>=20.1.0
